requirements.txt
telethon
python-dotenv
pandas
openpyxl
unicodedata2 # for some environments, for Amharic normalization 